+++ demo/src/main/AndroidManifest.xml
@@ -10,11 +10,10 @@
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
-        android:label="@string/app_name"
+        android:label="Engine-Braze Integration"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/AppTheme"
-        android:name=".MainApplication">
+        android:theme="@style/AppTheme">
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -22,7 +21,8 @@
             </intent-filter>
         </activity>
         <service android:name="com.factual.engine.FactualEngine"/>
-        <receiver android:name="com.factual.engine.braze.EngineBrazeBroadcastReceiver"/>
+        <receiver android:name="com.factual.engine.FactualClientReceiver"/>
+        <receiver android:name="com.factual.engine.braze.BrazeEngineBroadcastReceiver"/>
     </application>
 
 </manifest>
